bring back extensive password recovery message (#8296)
* bring back extensive password recovery message

* use Trans component inside useKratosT and fix some tooltips

* remove commented code

* remove more commented code
